requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
APScheduler==3.10.1
astropy==5.2.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
cachetools==5.3.0
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
coverage==7.2.2
cryptography==39.0.2
cssselect==1.2.0
cx-Oracle==8.3.0
cycler==0.11.0
dask==2023.3.1
decorator==5.1.1
desk==1.8.16
dnspython==2.3.0
docker==6.0.1
docutils==0.19
duckduckgo-search==2.8.5
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
flake8==6.0.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
gTTS==2.3.1
h5py==3.8.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.11.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.13.1
kiwisolver==1.4.4
locket==1.0.0
loguru==0.6.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.2
orjson==3.8.10
packaging==23.0
pandas==1.5.3
paramiko==3.1.0
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pkginfo==1.9.6
playsound==1.2.2
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyerfa==2.0.0.1
pyflakes==3.0.1
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.2.2
pytest-cov==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
readability-lxml==0.8.1
readme-renderer==37.3
redis==4.5.1
regex==2023.3.23
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.2
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4
Sphinx==6.1.3
sphinx-automodapi==0.15.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.6
stack-data==0.6.2
tiktoken==0.3.3
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
xlrd==2.0.1
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
